pyproject.toml
[tool.poetry]
name = "hw27"
version = "0.1.0"
description = ""
authors = ["aperushin <aperushin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
psycopg2 = "^2.9.6"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
drf-spectacular = "^0.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
